refactor(booking): four things were living in booking.service.ts; now four files
`booking.service.ts` sat at 971 lines against a 972 cap, and three separate
plans — booking-deposit (#20), scheduling-phase-b (#18), scheduling-phase-d
(#27) — all need to edit it and none of them could add a line. Pure move:
no query, message, status code, log key or ordering changed, and the 133
tests across tests/unit/{bookings,scheduling,calendar,usage} plus the
workers db-batch-atomicity spec pass untouched.

`fulfillBooking` alone was 513 of those lines, and it was four things.

`booking/booking-admission.ts` — EVERYTHING THAT MUST HOLD BEFORE A ROW
EXISTS: bot protection, widget origin allowlist, inspector-belongs-to-tenant,
booking-open, holiday block, and the slot claim. The seam is the absence of
writes. Every refusal here leaves the database exactly as it was; the moment
one row exists the failure mode changes from "refuse" to "compensate" (see
`arbitrateSlotRace`). Grouping them is also what makes it checkable that
quota is consumed after all of them — the property
`plan-quota-guarded-services.spec.ts` asserts.

`booking/booking-people.ts` — WHO the booking is for. Task 13 dropped the
legacy contact columns from `inspections`, so `inspection_people` is the only
persistence of who, and on this path it is written from two ends: the
referring agent resolved from `?ref=` before anything exists, the client
upserted after the rows commit. Eighty lines apart in the old file, which is
how one of them gets forgotten. Every write in the module is non-fatal by
design and the doc says so once, at the top, instead of five times inline.

`booking/booking-confirmation.ts` — what the world hears afterwards: the
Google Calendar push and the customer's confirmation email with its ICS
invite, inspector signature, credentials footer and SMS double-opt-in link.
The seam is "after the answer is sent" — the caller hands it to `waitUntil`,
so nothing in it can change the booker's response. `windowLabel` went with
it; the email is its only reader.

`booking/fulfill-booking.ts` keeps what remains, which is exactly the part
that writes and the part that cleans up after itself: the two row-creation
branches, the post-insert TOCTOU arbitration that may revoke what was just
written, and the scheduled-instant stamp over the survivors. Those three
cannot be separated.

`AvailabilityService` — a second exported class that shared nothing with the
first but the file — becomes `services/availability.service.ts`. It is the
WRITE side of availability (the weekly grid and its dated exceptions);
BookingService is the read side that turns those rows into bookable slots.
They never shared a query. Its three importers (`di.ts`, `types/hono.ts`,
`booking-delete-override-scope.spec.ts`) are repointed rather than given a
re-export.

TWO D1 HANDLES, AND THEY ARE NOT INTERCHANGEABLE. The original mixed
`drizzle(c.env.DB)` with `this.db` (`resolvePublicHolidayEffect`,
`new PeopleService({ DB: this.db })`), and unit tests construct the service
with a handle that is not `c.env.DB`. Both are explicit parameters now, with
the reason written down, because collapsing them to one is the obvious tidy
that would silently break those specs.

`originHeader` stays `string | undefined` rather than being normalised to
null on the way through the claim object: it is forwarded verbatim into
widget telemetry, where `undefined` omits the key and `null` writes one.

fulfillBooking stays a method — six specs and every caller reach it as
`c.var.services.booking.fulfillBooking(...)`.

Baseline ENTRY removed, not tightened. At 380 lines the file is governed by
the ordinary 400-line rule again; re-baselining at 380 would have rebuilt
the wall one line further out.

`type-check:app` run explicitly, not just `:api` — `types/hono.ts` is on the
hono/client RPC path and the api tsconfig does not build it.
