+++ server/services/booking.service.ts
@@ -1,35 +1,18 @@
 import type { Context } from 'hono';
 import { drizzle } from 'drizzle-orm/d1';
 import { eq, and, gte, lte, sql, inArray, isNull, ne } from 'drizzle-orm';
-import { availability, availabilityOverrides, calendarBlocks, inspections, inspectionInspectors, inspectionRequests, serviceInspectors, tenantConfigs, users, services as servicesTable, contactRoleProfiles, contacts } from '../lib/db/schema';
-import { CredentialService } from './credential.service';
-import { wallClockToEpochMs, resolveTenantTimeZone } from '../lib/tz';
-import { Errors } from '../lib/errors';
-import { safeISODate } from '../lib/date';
+import { availability, availabilityOverrides, calendarBlocks, inspections, inspectionInspectors, inspectionRequests, serviceInspectors, users } from '../lib/db/schema';
 import { logger } from '../lib/logger';
-import { fireAutomation } from './inspection/shared';
-import { normalizeLocale } from '../lib/i18n/contact-locale';
 import type { HonoConfig } from '../types/hono';
 import type { PublicBookingSchema } from '../lib/validations/booking.schema';
 import type { z } from '@hono/zod-openapi';
-import { createCalendarEvent } from '../api/calendar';
-import { loadOpenGoogleConnection } from '../lib/calendar/connection';
-import {
-    loadGoogleOAuthMode,
-    resolveGoogleOAuthCredentials,
-} from '../lib/calendar/resolve-google-oauth';
-import { canPushEvents } from '../lib/calendar/provider';
-import { getBookingHost, getBaseUrl } from '../lib/url';
-import { syncInspectionAssignments } from '../lib/db/assignment-links';
-import { PeopleService } from './people.service';
-import { INSPECTION_STATUS } from '../lib/status/inspection-status';
 import { buildSlotGrid } from '../lib/booking/slot-grid';
 import { loadSlotGridOptions } from '../lib/booking/slot-rules';
 import { computeBusyTimes } from '../lib/booking/busy-times';
 import { buildTenantSlotMap } from '../lib/booking/tenant-slot-map';
 import { resolvePublicHolidayEffect } from '../lib/holidays/load-tenant-holidays';
+import { fulfillBooking as runFulfillBooking } from './booking/fulfill-booking';
 import type { PlanQuotaGuard } from '../features/plan-quota/guard';
-
 /**
  * Service to handle public booking flow and availability lookups.
  */
@@ -380,592 +363,18 @@ export class BookingService {
      * enforcement and all fulfillment side effects, returning the same JSON
      * Response the handler used to return.
      */
+    /**
+     * Public booking fulfilment. The body lives in `./booking/fulfill-booking`
+     * and its three neighbours — admission, people, confirmation. It stays a
+     * method here because every caller and six specs reach it as
+     * `c.var.services.booking.fulfillBooking(...)`, and because the free
+     * function needs the handles this instance was constructed with.
+     */
     async fulfillBooking(
         c: Context<HonoConfig>,
         tenantId: string,
         body: z.infer<typeof PublicBookingSchema>,
     ) {
-        const service = c.var.services.booking;
-
-        // Bot Protection — always enforce when secret is configured
-        if (c.env.TURNSTILE_SECRET_KEY) {
-            if (!body.turnstileToken) throw Errors.Forbidden('Security verification token missing.');
-            const isValid = await service.verifyBotProtection(body.turnstileToken, c.env.TURNSTILE_SECRET_KEY);
-            if (!isValid) throw Errors.Forbidden('Security verification failed.');
-        }
-
-        // B2: when the booking originates from an embedded widget, enforce
-        // per-tenant origin allowlist. Non-embed (direct /book visit) submissions
-        // are unaffected.
-        const isWidgetSubmit = c.req.query('embed') === '1';
-        const originHeader = c.req.header('origin');
-        if (isWidgetSubmit) {
-            const ok = await c.var.services.widget.isOriginAllowed(tenantId, originHeader ?? null);
-            if (!ok) {
-                await c.var.services.widget.recordEvent(tenantId, 'error', { origin: originHeader, reason: 'origin_not_allowed' });
-                throw Errors.Forbidden('Widget submissions from this origin are not allowed for this workspace.');
-            }
-        }
-
-        const db = drizzle(c.env.DB);
-
-        // UC-A-1 — agent referral attribution. Resolve `?ref=<agentSlug>` (sent
-        // through the form as agentRefSlug) to a contacts.id in this tenant.
-        // Two requirements both need to hold:
-        //   1. A global agent user with that slug exists.
-        //   2. They have an `active` agent_tenant_links row for THIS tenant whose
-        //      inspectorContactId points at the agent's contact row.
-        // Either failure leaves referredByAgentId null — bookings with bad slugs
-        // still succeed; we just don't credit the (unknown) agent.
-        let resolvedAgentContactId: string | null = null;
-        if (body.agentRefSlug) {
-            try {
-                const agent = await db.select({ id: users.id })
-                    .from(users)
-                    .where(and(
-                        eq(users.slug, body.agentRefSlug),
-                        isNull(users.tenantId),
-                        eq(users.role, 'agent'),
-                    ))
-                    .get();
-                if (agent) {
-                    // IA-104 — the agent's contact in THIS tenant is the row
-                    // bound to their account; no link hop.
-                    const link = await db.select({ contactId: contacts.id })
-                        .from(contacts)
-                        .where(and(
-                            eq(contacts.agentUserId, agent.id),
-                            eq(contacts.tenantId, tenantId),
-                            isNull(contacts.agentRevokedAt),
-                        ))
-                        .get();
-                    resolvedAgentContactId = link?.contactId ?? null;
-                }
-            } catch (err) {
-                logger.warn('booking.agentRef.resolve.failed', {
-                    slug: body.agentRefSlug,
-                    tenantId,
-                    error: err instanceof Error ? err.message : String(err),
-                });
-            }
-        }
-
-        // IA-26 — inspectorId is now OPTIONAL. The company-level booking page
-        // submits without one (pure auto-assign); the legacy per-inspector
-        // deep link and the allowInspectorChoice dropdown still send it.
-        const serviceIdsForQual = (body.services ?? []).map(s => s.serviceId);
-        let inspectorId = body.inspectorId ?? null;
-
-        if (inspectorId) {
-            // B-16 — a supplied inspector must belong to the resolved tenant;
-            // a mismatched id (tampered payload or stale form) must not reach
-            // into another tenant's availability/inspection space.
-            const inspectorRow = await db.select({ id: users.id }).from(users)
-                .where(and(eq(users.id, inspectorId), eq(users.tenantId, tenantId)))
-                .get();
-            if (!inspectorRow) throw Errors.NotFound('Inspector not found.');
-        }
-
-        // B-16 (company-wide) — distinguish "nobody configured working hours"
-        // from a genuinely taken slot, with the honest not-open copy.
-        // qualifiedIds is computed once here and threaded through to avoid
-        // duplicate getQualifiedInspectorIds lookups in hasAnyHours / getTenantSlots.
-        const qualifiedIds = await service.getQualifiedInspectorIds(tenantId, serviceIdsForQual);
-        const bookingOpen = await service.hasAnyHours(tenantId, serviceIdsForQual, qualifiedIds);
-        if (!bookingOpen) {
-            throw Errors.Conflict('Online booking is not open yet. Please contact the company directly to schedule.');
-        }
-
-        const holiday = await resolvePublicHolidayEffect(this.db, tenantId, body.date);
-        if (holiday.effect === 'block') {
-            throw Errors.BadRequest(
-                holiday.name
-                    ? `The office is closed on ${holiday.name}. Please pick another date.`
-                    : 'The office is closed on this date. Please pick another date.',
-                'HOLIDAY_BLOCKED',
-            );
-        }
-
-        // Spec 3C / IA-26 — availability enforcement now runs on the tenant
-        // aggregation: a slot is bookable iff at least one QUALIFIED inspector
-        // is free (or the requested one, when the client chose).
-        let requestedTime: string;
-        switch (body.timeSlot) {
-            case 'morning':   requestedTime = '08:00'; break;
-            case 'afternoon': requestedTime = '13:00'; break;
-            case 'all-day':   requestedTime = '08:00'; break;
-            case 'custom':    requestedTime = body.customTime ?? '08:00'; break;
-        }
-        // KNOWN RACE (advisory check): the slot read and the inspection insert
-        // below are not atomic and D1 offers no row locks, so two concurrent
-        // submits for the last slot can both pass and double-book the same
-        // inspector (deterministic pickInspector converges on one person).
-        // Accepted for launch traffic; a post-insert recheck/compensation is
-        // tracked in the backlog. Do NOT "fix" by randomizing the pick — the
-        // determinism is intentional (idempotent re-submits).
-        const { slots } = await service.getTenantSlots(tenantId, body.date, serviceIdsForQual, qualifiedIds);
-        const target = slots.find(s => s.time === requestedTime);
-        const freeIds = (target?.inspectorIds ?? []).filter(id => !inspectorId || id === inspectorId);
-        if (freeIds.length === 0) {
-            throw Errors.Conflict('That time slot is no longer available. Please pick another time.');
-        }
-        if (!inspectorId) {
-            inspectorId = await service.pickInspector(tenantId, freeIds);
-            if (!inspectorId) throw Errors.Conflict('That time slot is no longer available. Please pick another time.');
-        }
-
-        // Sprint 2 S2-2 — When the customer selects multiple services, we route
-        // through InspectionRequestService so the resulting inspections are
-        // grouped under a parent request. The legacy single-service flow still
-        // creates a one-inspection request implicitly so dashboards can group
-        // every booking the same way.
-        const startIso = `${body.date}T${requestedTime}:00Z`;
-        const inspectionRequestService = c.var.services.inspectionRequest;
-        let createdRequestId: string;
-        let primaryInspectionId: string;
-        let allInspectionIds: string[] = [];
-        // Task 7b (people-role-profiles) — set only by the direct-insert
-        // (legacy single-service) branch below. The multi-service branch
-        // routes through InspectionRequestService.create, which owns its
-        // own inspection_people write for the inspections it creates.
-        let directInsertInspectionId: string | null = null;
-        // Booked duration from the chosen service(s); NULL when the legacy path
-        // carries no explicit service (falls back to the time-slot window below).
-        let bookedServiceDurationMin: number | null = null;
-
-        if (body.services && body.services.length > 0) {
-            const serviceIds = body.services.map(s => s.serviceId);
-            const svcRows = await db.select().from(servicesTable)
-                .where(and(eq(servicesTable.tenantId, tenantId), inArray(servicesTable.id, serviceIds)))
-                .all();
-            if (svcRows.length !== serviceIds.length) {
-                throw Errors.BadRequest('One or more services were not found.');
-            }
-            // Total booked minutes across the selected services (back-to-back);
-            // NULL when none carry a duration, so the time-slot window is used.
-            bookedServiceDurationMin =
-                svcRows.reduce((sum, s) => sum + (s.durationMinutes ?? 0), 0) || null;
-            const subs = svcRows.map(s => {
-                const sub: { templateId: string; price: number } = {
-                    templateId: s.templateId ?? '',
-                    price:      s.price ?? 0,
-                };
-                if (!sub.templateId) throw Errors.BadRequest(`Service '${s.name}' has no template configured.`);
-                return sub;
-            });
-            const created = await inspectionRequestService.create(tenantId, {
-                clientName:      body.clientName,
-                clientEmail:     body.clientEmail,
-                propertyAddress: body.address,
-                scheduledAt:     startIso,
-                inspectorId,
-                referredByAgentId: resolvedAgentContactId,
-            }, subs);
-            createdRequestId = created.id;
-            allInspectionIds = created.inspections.map(i => i.id);
-            primaryInspectionId = allInspectionIds[0] ?? '';
-        } else {
-            primaryInspectionId = crypto.randomUUID();
-            createdRequestId = `req-${primaryInspectionId}`;
-            const now = new Date();
-            // Quota is consumed AFTER every precondition check above (bot
-            // protection, widget origin, inspector ownership, booking-open,
-            // slot availability) and BEFORE either row below is inserted —
-            // the request row must never be orphaned (created with no
-            // inspection behind it) because the tenant hit the cap.
-            await this.planQuota?.consumeInspection(tenantId);
-            // Insert one-inspection request first so the FK is satisfied.
-            await db.insert(inspectionRequests).values({
-                id:              createdRequestId,
-                tenantId,
-                clientName:      body.clientName,
-                clientEmail:     body.clientEmail,
-                propertyAddress: body.address,
-                scheduledAt:     new Date(startIso),
-                status:          'pending',
-                totalAmount:     0,
-                paymentStatus:   'unpaid',
-                createdAt:       now,
-                updatedAt:       now,
-            });
-            await db.insert(inspections).values({
-                id: primaryInspectionId,
-                tenantId,
-                inspectorId,
-                propertyAddress: body.address,
-                // B-28 adjacent fix — store the full start ISO like the
-                // multi-service path (inspection-request.service create) does.
-                // Busy checks read HH:MM at slice(11,16) of this value; the old
-                // bare `body.date` never marked the slot busy, so even
-                // sequential double-booking succeeded.
-                date: startIso,
-                status: INSPECTION_STATUS.REQUESTED,
-                paymentStatus: 'unpaid',
-                price: 0,
-                requestId: createdRequestId,
-                createdAt: now
-            });
-            // DB-8: mirror assignment into inspection_inspectors link table.
-            // Non-fatal — the link table is a denormalized mirror; a sync failure
-            // must never 500 an anonymous booker whose inspection row already committed.
-            try {
-                await syncInspectionAssignments(db, tenantId, primaryInspectionId, { inspectorId });
-            } catch (e) {
-                logger.error('booking.assignment-sync.failed', { inspectionId: primaryInspectionId }, e instanceof Error ? e : undefined);
-            }
-            allInspectionIds = [primaryInspectionId];
-            directInsertInspectionId = primaryInspectionId;
-        }
-        const inspectionId = primaryInspectionId;
-
-        // B-28 — post-insert TOCTOU recheck. Runs after our insert and BEFORE
-        // any side effect (confirmation email, calendar event, notifications)
-        // so a losing booker only ever sees the 409, never a confirmation for
-        // a booking that then vanishes. The arbitration is deterministic
-        // (earliest (createdAt, id) wins), so of two racers exactly one
-        // self-compensates here while the other proceeds untouched.
-        const verdict = await service.arbitrateSlotRace(
-            tenantId, inspectorId!, body.date, requestedTime, createdRequestId,
-        );
-        if (verdict === 'lose') {
-            await service.revokeBooking(tenantId, createdRequestId);
-            throw Errors.Conflict('That time slot is no longer available. Please pick another time.');
-        }
-
-        // A-polish 9b — stamp the precise scheduled instant on every inspection
-        // this booking created. The wall-clock slot time is interpreted in the
-        // TENANT tz (not the naive :00Z of the startIso busy-check key), so
-        // downstream conflict detection, Google push, and the ICS feed reason
-        // about real instants. inspections.date is deliberately left untouched —
-        // it still keys the HH:MM busy-checks via slice(11,16). Non-fatal: the
-        // inspection rows already committed, so a stamp failure must not 500 the
-        // booker (conflict detection just falls back to the hour-bucket).
-        const tzRow = await db.select({ defaultTimezone: tenantConfigs.defaultTimezone })
-            .from(tenantConfigs).where(eq(tenantConfigs.tenantId, tenantId)).get();
-        const tenantTz = resolveTenantTimeZone(tzRow?.defaultTimezone);
-        const scheduledStartMs = wallClockToEpochMs(body.date, requestedTime, tenantTz);
-        const slotWindowMin =
-            body.timeSlot === 'all-day' ? 540
-            : body.timeSlot === 'morning' || body.timeSlot === 'afternoon' ? 240
-            : 180;
-        const durationMin = bookedServiceDurationMin ?? slotWindowMin;
-        const scheduledEndMs = scheduledStartMs + durationMin * 60000;
-        try {
-            await db.update(inspections)
-                .set({
-                    scheduledStartMs: new Date(scheduledStartMs),
-                    scheduledEndMs: new Date(scheduledEndMs),
-                    durationMin,
-                })
-                .where(and(
-                    inArray(inspections.id, allInspectionIds),
-                    eq(inspections.tenantId, tenantId),
-                ));
-        } catch (e) {
-            logger.warn('booking.scheduled-instant.stamp.failed', {
-                inspectionIds: allInspectionIds,
-                error: e instanceof Error ? e.message : String(e),
-            });
-        }
-
-        // IA-18 (#111) — capture the booker as a Client contact and link it to
-        // ALL inspections this booking created so their client appears in
-        // Contacts and on the inspector portal People card.
-        //
-        // Placement: AFTER arbitration. A losing booker self-revokes and throws
-        // above, so we never stamp a contact onto inspections that were just
-        // deleted. (A stray contact row is harmless on its own — what we avoid
-        // is a clientContactId pointing at vanished inspections.) It also runs
-        // BEFORE the side-effect block to keep the synchronous DB writes
-        // grouped before async waitUntil work.
-        //
-        // Non-fatal: a booking must NEVER fail because of contact bookkeeping.
-        // Any error is logged (NO client email — only inspection ids + message)
-        // and swallowed; the inspection rows already committed regardless.
-        let bookingClientContactId: string | null = null;
-        if (body.clientEmail || body.clientName) {
-            try {
-                const { id: clientContactId } = await c.var.services.contact.upsertClientContact(tenantId, {
-                    name:  body.clientName,
-                    email: body.clientEmail,
-                    type:  'client',
-                    // Reduced to a locale we actually have messages for, so a
-                    // regional variant lands on its catalogue and anything we
-                    // cannot speak is stored as NULL rather than as a promise
-                    // we would break at send time.
-                    locale: normalizeLocale(body.locale),
-                });
-                bookingClientContactId = clientContactId;
-            } catch (e) {
-                logger.warn('booking.client-contact.upsert.failed', {
-                    inspectionIds: allInspectionIds,
-                    error: e instanceof Error ? e.message : String(e),
-                });
-            }
-        }
-
-        // Task 7b (people-role-profiles), FIXED — mirror client + buyer_agent
-        // into inspection_people. Task 13 dropped the legacy clientContactId /
-        // referredByAgentId columns from inspections, so this is now the ONLY
-        // persistence of WHO. Client covers EVERY allInspectionIds entry
-        // (bookingClientContactId is linked to all of them, incl.
-        // multi-service sub-inspections, which only get buyer_agent from
-        // InspectionRequestService.create — the original bug wrongly scoped
-        // the client write to directInsertInspectionId alone). Non-fatal.
-        if (bookingClientContactId || (directInsertInspectionId && resolvedAgentContactId)) {
-            try {
-                const roleRows = await db.select({ id: contactRoleProfiles.id, key: contactRoleProfiles.key })
-                    .from(contactRoleProfiles)
-                    .where(and(eq(contactRoleProfiles.tenantId, tenantId), eq(contactRoleProfiles.active, true)));
-                const roleIdByKey = new Map(roleRows.map(r => [r.key, r.id]));
-                const people = new PeopleService({ DB: this.db });
-                const clientRoleId = roleIdByKey.get('client');
-                if (bookingClientContactId && clientRoleId) {
-                    for (const inspId of allInspectionIds) {
-                        await people.addPerson(tenantId, inspId, bookingClientContactId, clientRoleId);
-                    }
-                }
-                const buyerAgentRoleId = roleIdByKey.get('buyer_agent');
-                if (directInsertInspectionId && resolvedAgentContactId && buyerAgentRoleId) {
-                    await people.addPerson(tenantId, directInsertInspectionId, resolvedAgentContactId, buyerAgentRoleId);
-                }
-            } catch (err) {
-                logger.error('inspection-people write from booking create failed', { inspectionIds: allInspectionIds }, err instanceof Error ? err : undefined);
-            }
-        }
-
-        // Track L (D6, path A) — self-book SMS opt-in. The checkbox is unchecked by
-        // default; when ticked we record a `granted` consent event (captured_via=
-        // booking_form) keyed on the client contact. Non-fatal: a consent write must
-        // never fail the booking (the inspection rows already committed).
-        if (body.smsOptin && bookingClientContactId) {
-            try {
-                const { SmsConsentService } = await import('./sms-consent.service');
-                await new SmsConsentService(c.env.DB).record(
-                    tenantId, bookingClientContactId, 'granted', 'booking_form',
-                    { ip: c.req.header('CF-Connecting-IP'), userAgent: c.req.header('User-Agent') },
-                );
-            } catch (e) {
-                logger.warn('booking.sms-optin.record.failed', {
-                    inspectionId, error: e instanceof Error ? e.message : String(e),
-                });
-            }
-        }
-
-        // Sprint 1 C-6 — map window option to a human-readable label for the
-        // calendar event + confirmation email.
-        const windowLabel: Record<typeof body.timeSlot, string> = {
-            'morning':   'Morning (8:00 AM – 12:00 PM)',
-            'afternoon': 'Afternoon (12:00 PM – 4:00 PM)',
-            'all-day':   'All day (8:00 AM – 5:00 PM)',
-            'custom':    body.customTime ? `${body.customTime}` : 'Custom time',
-        };
-
-        // Async tasks
-        c.executionCtx.waitUntil((async () => {
-            const inspector = await db.select().from(users).where(eq(users.id, inspectorId!)).get();
-            const open = await loadOpenGoogleConnection(
-                c.env.DB,
-                tenantId,
-                inspectorId!,
-                c.env.JWT_SECRET,
-                c.env.JWT_SECRET_PREVIOUS,
-            );
-            if (open && canPushEvents(open.connection.capabilities)) {
-                const oauthMode = await loadGoogleOAuthMode(c.env.DB, tenantId);
-                const oauthCreds = await resolveGoogleOAuthCredentials(c.env, tenantId, oauthMode);
-                if (oauthCreds) {
-                    const startDateTime = `${body.date}T${requestedTime}:00Z`;
-                    await createCalendarEvent(
-                        oauthCreds.clientId,
-                        oauthCreds.clientSecret,
-                        open.credentials.refreshToken,
-                        open.connection.calendarId,
-                        `Inspection: ${body.address}`,
-                        startDateTime,
-                        body.address,
-                    ).catch(e => logger.error('Calendar sync failed', {}, e instanceof Error ? e : undefined));
-                }
-            }
-
-            const emailService = c.var.services.email;
-
-            // Sprint 1 C-10 — build the ICS event so the confirmation email
-            // carries a calendar invite the customer can import into Apple
-            // Calendar / Google Calendar. Duration defaults to 3 hours, with
-            // 4 hours for morning/afternoon windows and 9 hours for all-day.
-            const startMs = new Date(`${body.date}T${requestedTime}:00Z`).getTime();
-            let durationHours: number;
-            switch (body.timeSlot) {
-                case 'all-day':   durationHours = 9; break;
-                case 'morning':
-                case 'afternoon': durationHours = 4; break;
-                default:          durationHours = 3; break;
-            }
-            const endMs = startMs + durationHours * 60 * 60 * 1000;
-            // Booking-confirmation greeting falls back to the brand, never the
-            // inspector's inbox — keeps the email looking professional even if a
-            // legacy account is missing a display name.
-            const inspectorName = inspector?.name || c.env.APP_NAME || 'Your inspector';
-            const inspectorEmail = inspector?.email || c.env.SENDER_EMAIL || `noreply@${c.env.APP_NAME?.toLowerCase().replace(/\s/g, '') || 'inspector'}.com`;
-
-            // Spec B — the assigned inspector's active credentials, for the footer.
-            // Via the shared mapper, so this footer and the email signature can
-            // never disagree about the badge URL form.
-            const bookingCreds = inspector
-                ? await new CredentialService(c.env.DB).listRenderable(tenantId, inspectorId!)
-                : [];
-            // Sprint B-4a — append inspector signature so customers can rebook
-            // with the same inspector via the per-inspector booking link.
-            const sigInspector = inspector ? {
-                name:          inspector.name ?? null,
-                email:         inspector.email ?? null,
-                phone:         inspector.phone ?? null,
-                slug:          inspector.slug ?? null,
-                credentials:   bookingCreds,
-            } : undefined;
-            // Track L (D6, path B) — double-opt-in link injected at the RENDERER
-            // level (not gated on any automation rule) so disabling a rule never
-            // removes the only opt-in path. The token self-describes (tenant,
-            // contact) — see lib/sms/optin-token.ts. Best-effort: a token failure
-            // simply omits the link.
-            let smsOptinUrl: string | undefined;
-            if (bookingClientContactId && c.env.JWT_SECRET) {
-                try {
-                    const { mintOptinToken } = await import('../lib/sms/optin-token');
-                    const token = await mintOptinToken(tenantId, bookingClientContactId, c.env.JWT_SECRET);
-                    smsOptinUrl = `${getBaseUrl(c)}/sms-optin/${encodeURIComponent(token)}`;
-                } catch (e) {
-                    logger.warn('booking.sms-optin.mint.failed', { inspectionId, error: e instanceof Error ? e.message : String(e) });
-                }
-            }
-
-            await emailService.sendBookingConfirmation(
-                body.clientEmail,
-                body.clientName,
-                body.address,
-                body.date,
-                windowLabel[body.timeSlot],
-                {
-                    uid:            `inspection-${inspectionId}`,
-                    summary:        `Home Inspection at ${body.address}`,
-                    description:    `Inspector: ${inspectorName}\nWindow: ${windowLabel[body.timeSlot]}\n\nWe will send your detailed report within 24 hours of completion.`,
-                    location:       body.address,
-                    start:          new Date(startMs),
-                    end:            new Date(endMs),
-                    organizerEmail: inspectorEmail,
-                    organizerName:  inspectorName,
-                },
-                sigInspector,
-                getBookingHost(c),
-                smsOptinUrl,
-            ).catch(e => logger.error('Booking confirmation email failed', {}, e instanceof Error ? e : undefined));
-        })());
-
-        if (isWidgetSubmit) {
-            c.executionCtx.waitUntil(
-                c.var.services.widget.recordEvent(tenantId, 'success', { origin: originHeader, inspectionId })
-            );
-        }
-
-        // B3 — the office alert is a rule now (`Office alert — new booking`,
-        // recipientKind 'staff', channel in_app). `booking.received` is its own
-        // trigger rather than a reuse of `inspection.created`: a booking is a
-        // stranger arriving through the public form, while an inspection can
-        // also be created by the office itself, and alerting someone about
-        // their own action is noise.
-        c.executionCtx.waitUntil(
-            fireAutomation(c.env.DB, tenantId, inspectionId, 'booking.received'),
-        );
-
-        return c.json({
-            success: true,
-            data: {
-                success: true,
-                inspectionId,
-                requestId: createdRequestId,
-                inspectionIds: allInspectionIds,
-            }
-        }, 200);
-    }
-}
-
-/**
- * Service to manage internal inspector availability schedules.
- */
-export class AvailabilityService {
-    constructor(private db: D1Database) {}
-
-    private getDrizzle() {
-        return drizzle(this.db);
-    }
-
-    /**
-     * Replaces the entire weekly schedule for an inspector.
-     */
-    async updateWeeklySchedule(tenantId: string, inspectorId: string, slots: { dayOfWeek: number; startTime: string; endTime: string }[]) {
-        const db = this.getDrizzle();
-        
-        await db.delete(availability).where(and(
-            eq(availability.tenantId, tenantId),
-            eq(availability.inspectorId, inspectorId)
-        ));
-
-        if (slots.length > 0) {
-            await db.insert(availability).values(
-                slots.map(s => ({
-                    id: crypto.randomUUID(),
-                    tenantId,
-                    inspectorId,
-                    dayOfWeek: s.dayOfWeek,
-                    startTime: s.startTime,
-                    endTime: s.endTime,
-                    createdAt: new Date(),
-                }))
-            );
-        }
-    }
-
-    /**
-     * Adds a specific availability override.
-     */
-    async addOverride(tenantId: string, data: {
-        inspectorId: string;
-        date: string;
-        isAvailable: boolean;
-        startTime?: string | null | undefined;
-        endTime?: string | null | undefined;
-    }) {
-        const db = this.getDrizzle();
-        const newOverride = {
-            id: crypto.randomUUID(),
-            tenantId,
-            inspectorId: data.inspectorId,
-            date: data.date,
-            isAvailable: data.isAvailable,
-            startTime: data.startTime || null,
-            endTime: data.endTime || null,
-            createdAt: new Date(),
-        };
-
-        await db.insert(availabilityOverrides).values(newOverride);
-        return {
-            ...newOverride,
-            createdAt: safeISODate(newOverride.createdAt)
-        };
-    }
-
-    /**
-     * Deletes an availability override.
-     */
-    async deleteOverride(tenantId: string, id: string) {
-        const db = this.getDrizzle();
-        const existing = await db.select().from(availabilityOverrides).where(and(
-            eq(availabilityOverrides.id, id),
-            eq(availabilityOverrides.tenantId, tenantId)
-        )).get();
-
-        if (!existing) throw Errors.NotFound('Override not found');
-        await db.delete(availabilityOverrides).where(and(eq(availabilityOverrides.id, id), eq(availabilityOverrides.tenantId, tenantId)));
+        return runFulfillBooking({ d1: this.db, planQuota: this.planQuota }, c, tenantId, body);
     }
 }
