+++ server/types/hono.ts
@@ -235,7 +235,8 @@ import type { AdminService } from '../services/admin.service';
 import type { AIService } from '../services/ai.service';
 import type { AuthService } from '../services/auth.service';
 import type { UserSyncOutbox } from '../lib/integration/user-sync';
-import type { BookingService, AvailabilityService } from '../services/booking.service';
+import type { BookingService } from '../services/booking.service';
+import type { AvailabilityService } from '../services/availability.service';
 import type { BrandingService } from '../services/branding.service';
 import type { LegalVersionService } from '../services/legal-version.service';
 import type { EmailService } from '../services/email.service';
